Added program_agencies method to GetPrograms class
This method fetches the JSON data of after-school programs using the get_programs method.

It then iterates over the fetched data, extracting the 'agency' field from each program.

The extracted agencies are appended to a list, which is returned by the method.

This provides a convenient way to get a list of all agencies running after-school programs.